+++ src/extension.ts
@@ -1,36 +1,36 @@
 import * as vscode from "vscode";
 
-import { createCommand } from "./commands";
 import { RefactoringActionProvider } from "./action-providers";
+import { createCommand } from "./commands";
 import { Refactoring, RefactoringWithActionProvider } from "./types";
 
 import addNumericSeparator from "./refactorings/add-numeric-separator";
 import changeSignature from "./refactorings/change-signature";
-import convertForToForEach from "./refactorings/convert-for-to-for-each";
 import convertForEachToForOf from "./refactorings/convert-for-each-to-for-of";
+import convertForToForEach from "./refactorings/convert-for-to-for-each";
 import convertIfElseToSwitch from "./refactorings/convert-if-else-to-switch";
-import convertSwitchToIfElse from "./refactorings/convert-switch-to-if-else";
 import convertIfElseToTernary from "./refactorings/convert-if-else-to-ternary";
+import convertLetToConst from "./refactorings/convert-let-to-const";
+import convertSwitchToIfElse from "./refactorings/convert-switch-to-if-else";
 import convertTernaryToIfElse from "./refactorings/convert-ternary-to-if-else";
 import convertToArrowFunction from "./refactorings/convert-to-arrow-function";
 import convertToTemplateLiteral from "./refactorings/convert-to-template-literal";
-import convertLetToConst from "./refactorings/convert-let-to-const";
 import createFactoryForConstructor from "./refactorings/create-factory-for-constructor";
 import destructureObject from "./refactorings/destructure-object";
 import extract from "./refactorings/extract";
 import extractGenericType from "./refactorings/extract-generic-type";
 import extractInterface from "./refactorings/extract-interface";
 import flipIfElse from "./refactorings/flip-if-else";
+import flipOperator from "./refactorings/flip-operator";
 import flipTernary from "./refactorings/flip-ternary";
-import flipYodaCondition from "./refactorings/flip-yoda-condition";
 import inline from "./refactorings/inline";
+import invertBooleanLogic from "./refactorings/invert-boolean-logic";
 import liftUpConditional from "./refactorings/lift-up-conditional";
 import mergeIfStatements from "./refactorings/merge-if-statements";
 import mergeWithPreviousIfStatement from "./refactorings/merge-with-previous-if-statement";
 import moveStatementDown from "./refactorings/move-statement-down";
 import moveStatementUp from "./refactorings/move-statement-up";
 import moveToExistingFile from "./refactorings/move-to-existing-file";
-import invertBooleanLogic from "./refactorings/invert-boolean-logic";
 import reactConvertToPureComponent from "./refactorings/react/convert-to-pure-component";
 import reactExtractUseCallback from "./refactorings/react/extract-use-callback";
 import removeDeadCode from "./refactorings/remove-dead-code";
@@ -99,7 +99,7 @@ const refactorings: { [key: string]: ConfiguredRefactoring } = {
       createFactoryForConstructor,
       flipIfElse,
       flipTernary,
-      flipYodaCondition,
+      flipOperator,
       inline,
       liftUpConditional,
       mergeIfStatements,